** Major Changes **
Replaced the OpenCV build java libraries with the JavaCV wrappers and
presets for the C++ version of OpenCV.

--- The stratagy for analysing video has now changed. Instead of doing
all post processing of individual frames of a video, we are doing real
time analysis on a video. ---

Added a VideoMatriarch to hold the video once it has been selected.
Added a LoadedVideo and NullVideo to wrap the JavaCV FrameGrabber, and
act as our main model object.
Added a VideoFrameIndex class to use in the VideoInfoMatriarch's map.
This will probably be removed in the future.

Signed-off-by: Theo Chupp <chuppthe@msu.edu>